refactor(milli): move facet range search tests to a separate file
`search/facet/facet_range_search.rs` carried ~350 lines of tests at the
end of the file, which slows down rust-analyzer and makes the module
harder to navigate.

Move the whole `#[cfg(test)] mod tests` block to a sibling
`facet_range_search_test.rs` file, declared in `search/facet/mod.rs` via
`#[cfg(test)] mod facet_range_search_test;` (following the convention
used for `autobatcher_test.rs`). The single `use super::...` import is
updated to `use super::facet_range_search::...` to account for the new
parent module. The `milli_snap` snapshots are renamed from
`snapshots/facet_range_search.rs/` to
`snapshots/facet_range_search_test.rs/` to match the new source file
name. No test logic is changed.

Part of #5116
