+++ crates/milli/src/search/facet/mod.rs
@@ -16,6 +16,8 @@ use crate::{Index, Result};
 mod facet_distribution;
 mod facet_distribution_iter;
 mod facet_range_search;
+#[cfg(test)]
+mod facet_range_search_test;
 mod facet_sort_ascending;
 mod facet_sort_descending;
 mod filter;
